fix(arch): P0 defects — SSOT quota type, tokenType constants, wire companion
P0-1: settings.ts PlatformType now imports from canonical
gatewayPlatforms.ts instead of maintaining an independent stale union;
SettingsView.vue v-for loops use ALLOWED_QUOTA_PLATFORMS; drift checker
gains CHECK 12 (Go AllowedQuotaPlatforms ↔ TS ALLOWED_QUOTA_PLATFORMS).

P0-2: Replace 25 hardcoded tokenType string literals ("oauth",
"apikey", "bedrock", "service_account") with domain.AccountType*
constants across gateway_service.go and openai_gateway_service.go.

P0-3: Extract 9 ProvideTK* wire providers + 6 sentinel types (288
lines) from wire.go into wire_tk.go companion; update 3 sentinel
registry entries in gateway-tk.json to point to new file.
