+++ scripts/checks/test_platform_registry_drift.py
@@ -16,6 +16,7 @@ _spec.loader.exec_module(prd)
 
 
 PLATFORMS = ["anthropic", "gemini", "openai", "antigravity", "newapi", "kiro", "grok"]
+QUOTA_PLATFORMS = ["anthropic", "openai", "gemini", "antigravity", "grok"]
 ACCOUNT_TYPES = ["oauth", "setup-token", "apikey", "upstream", "bedrock", "service_account"]
 ROLES = ["admin", "user"]
 REDEEM_TYPES = ["balance", "concurrency", "subscription", "invitation"]
@@ -205,10 +206,26 @@ def _fixture(
     )
     _write(
         root,
+        "backend/internal/service/domain_constants.go",
+        """
+        package service
+
+        var AllowedQuotaPlatforms = []string{
+            PlatformAnthropic,
+            PlatformOpenAI,
+            PlatformGemini,
+            PlatformAntigravity,
+            PlatformGrok,
+        }
+        """,
+    )
+    _write(
+        root,
         "frontend/src/constants/gatewayPlatforms.ts",
         f"""
         export const OPENAI_COMPAT_PLATFORMS = {_ts_array(["openai", "newapi", "grok"])} as const
         export const GROUP_DISPATCH_CONFIG_PLATFORMS = {_ts_array(["openai", "newapi", "gemini", "grok"])} as const
+        export const ALLOWED_QUOTA_PLATFORMS = {_ts_array(QUOTA_PLATFORMS)} as const
 
         const SOFT_BADGE: Record<string, string> = {_ts_record(soft_badge or PLATFORMS)}
